Split up nfo writer
