+++ src/main/java/kodi/NfoWriter.java
@@ -1,0 +1,110 @@
+package kodi;
+
+import lombok.val;
+import org.dom4j.Document;
+import org.dom4j.DocumentException;
+import org.dom4j.DocumentHelper;
+import org.dom4j.io.OutputFormat;
+import org.dom4j.io.XMLWriter;
+
+import javax.xml.namespace.QName;
+import javax.xml.stream.XMLEventFactory;
+import javax.xml.stream.XMLEventWriter;
+import javax.xml.stream.XMLStreamException;
+import javax.xml.stream.events.Attribute;
+import java.io.IOException;
+import java.io.StringWriter;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.StandardOpenOption;
+import java.util.List;
+
+public class NfoWriter {
+    protected XMLEventWriter writer;
+    protected XMLEventFactory factory;
+
+    protected static void prettyPrint(String content, Path file) throws IOException, DocumentException {
+        OutputFormat format = OutputFormat.createPrettyPrint();
+        format.setIndentSize(4);
+        format.setSuppressDeclaration(false);
+        format.setEncoding(StandardCharsets.UTF_8.displayName());
+
+        Document document = DocumentHelper.parseText(content);
+        StringWriter sw = new StringWriter();
+        XMLWriter writer = new XMLWriter(sw, format);
+        writer.write(document);
+
+        try (val fileWriter = Files.newBufferedWriter(file, StandardOpenOption.WRITE, StandardOpenOption.TRUNCATE_EXISTING, StandardOpenOption.CREATE)) {
+            fileWriter.write(sw.toString());
+        }
+    }
+
+    protected List<Attribute> attributes(String name, String value) {
+        return List.of(attribute(name, value));
+    }
+
+    protected Attribute attribute(String name, String value) {
+        return factory.createAttribute(name, value);
+    }
+
+    protected Attribute attribute(String name, int value) {
+        return attribute(name, String.valueOf(value));
+    }
+
+    protected void startDocument() throws XMLStreamException {
+        writer.add(factory.createStartDocument("UTF-8", "1.0", true));
+    }
+
+    protected void writeTag(String tag, SeriesNfoWriter.IContentWriter content) throws XMLStreamException {
+        writer.add(factory.createStartElement(QName.valueOf(tag), null, null));
+        content.write();
+        writer.add(factory.createEndElement(QName.valueOf(tag), null));
+    }
+
+    protected void writeTag(String tag, int content) throws XMLStreamException {
+        writer.add(factory.createStartElement(QName.valueOf(tag), null, null));
+        writer.add(factory.createCharacters(String.valueOf(content)));
+        writer.add(factory.createEndElement(QName.valueOf(tag), null));
+    }
+
+    protected void writeTag(String tag, long content) throws XMLStreamException {
+        writer.add(factory.createStartElement(QName.valueOf(tag), null, null));
+        writer.add(factory.createCharacters(String.valueOf(content)));
+        writer.add(factory.createEndElement(QName.valueOf(tag), null));
+    }
+
+    protected void writeTag(String tag, long content, List<Attribute> attributes) throws XMLStreamException {
+        writer.add(factory.createStartElement(QName.valueOf(tag), attributes.iterator(), null));
+        writer.add(factory.createCharacters(String.valueOf(content)));
+        writer.add(factory.createEndElement(QName.valueOf(tag), null));
+    }
+
+    protected void writeTag(String tag, double content) throws XMLStreamException {
+        writer.add(factory.createStartElement(QName.valueOf(tag), null, null));
+        writer.add(factory.createCharacters(String.valueOf(content)));
+        writer.add(factory.createEndElement(QName.valueOf(tag), null));
+    }
+
+    protected void writeTag(String tag, String content) throws XMLStreamException {
+        writer.add(factory.createStartElement(QName.valueOf(tag), null, null));
+        writer.add(factory.createCharacters(content));
+        writer.add(factory.createEndElement(QName.valueOf(tag), null));
+    }
+
+    protected void writeTag(String tag, String content, List<Attribute> attributes) throws XMLStreamException {
+        writer.add(factory.createStartElement(QName.valueOf(tag), attributes.iterator(), null));
+        writer.add(factory.createCharacters(content));
+        writer.add(factory.createEndElement(QName.valueOf(tag), null));
+    }
+
+    protected void writeTag(String tag, List<Attribute> attributes, SeriesNfoWriter.IContentWriter content) throws XMLStreamException {
+        writer.add(factory.createStartElement(QName.valueOf(tag), attributes.iterator(), null));
+        content.write();
+        writer.add(factory.createEndElement(QName.valueOf(tag), null));
+    }
+
+    interface IContentWriter {
+        void write() throws XMLStreamException;
+    }
+}
