+++ src/main/java/kodi/SeriesNfoWriter.java
@@ -5,7 +5,6 @@ import kodi.nfo.Actor;
 import kodi.nfo.Artwork;
 import kodi.nfo.Episode;
 import kodi.nfo.Series;
-import kodi.tvdb.TvDbArtworksResponse;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.java.Log;
@@ -15,7 +14,6 @@ import org.dom4j.DocumentException;
 import org.dom4j.DocumentHelper;
 import org.dom4j.io.OutputFormat;
 import org.dom4j.io.XMLWriter;
-import processing.FileInfo;
 
 import javax.xml.namespace.QName;
 import javax.xml.stream.*;
@@ -34,14 +32,11 @@ import java.util.List;
 
 @Log
 @RequiredArgsConstructor(staticName = "forSeries")
-public class NfoGenerator {
+public class SeriesNfoWriter extends NfoWriter {
     private static final String seriesNfo = "tvshow.nfo";
     @Getter
     final Series series;
     Episode episode;
-
-    private XMLEventWriter writer;
-    private XMLEventFactory factory;
 
     public void writeNfoFiles(Episode episode, boolean overwriteSeries, boolean overwriteEpisode) {
         this.episode = episode;
@@ -64,27 +59,6 @@ public class NfoGenerator {
             }
         } catch (XMLStreamException | IOException | DocumentException e) {
             throw new RuntimeException(e);
-        }
-    }
-
-    private String getEpisodeFileName(Path episodePath) {
-        val rootFileName = episodePath.toString();
-        return rootFileName.substring(0, rootFileName.lastIndexOf("."));
-    }
-
-    private void prettyPrint(String content, Path file) throws IOException, DocumentException {
-        OutputFormat format = OutputFormat.createPrettyPrint();
-        format.setIndentSize(4);
-        format.setSuppressDeclaration(false);
-        format.setEncoding(StandardCharsets.UTF_8.displayName());
-
-        Document document = DocumentHelper.parseText(content);
-        StringWriter sw = new StringWriter();
-        XMLWriter writer = new XMLWriter(sw, format);
-        writer.write(document);
-
-        try (val fileWriter = Files.newBufferedWriter(file, StandardOpenOption.WRITE, StandardOpenOption.TRUNCATE_EXISTING, StandardOpenOption.CREATE)) {
-            fileWriter.write(sw.toString());
         }
     }
 
@@ -245,74 +219,5 @@ public class NfoGenerator {
                 writeTag("order", actor.getOrder());
             });
         }
-    }
-
-
-    private List<Attribute> attributes(String name, String value) {
-        return List.of(attribute(name, value));
-    }
-
-    private Attribute attribute(String name, String value) {
-        return factory.createAttribute(name, value);
-    }
-
-    private Attribute attribute(String name, int value) {
-        return attribute(name, String.valueOf(value));
-    }
-
-    private void startDocument() throws XMLStreamException {
-        writer.add(factory.createStartDocument("UTF-8", "1.0", true));
-    }
-
-    private void writeTag(String tag, IContentWriter content) throws XMLStreamException {
-        writer.add(factory.createStartElement(QName.valueOf(tag), null, null));
-        content.write();
-        writer.add(factory.createEndElement(QName.valueOf(tag), null));
-    }
-
-    private void writeTag(String tag, int content) throws XMLStreamException {
-        writer.add(factory.createStartElement(QName.valueOf(tag), null, null));
-        writer.add(factory.createCharacters(String.valueOf(content)));
-        writer.add(factory.createEndElement(QName.valueOf(tag), null));
-    }
-
-    private void writeTag(String tag, long content) throws XMLStreamException {
-        writer.add(factory.createStartElement(QName.valueOf(tag), null, null));
-        writer.add(factory.createCharacters(String.valueOf(content)));
-        writer.add(factory.createEndElement(QName.valueOf(tag), null));
-    }
-
-    private void writeTag(String tag, long content, List<Attribute> attributes) throws XMLStreamException {
-        writer.add(factory.createStartElement(QName.valueOf(tag), attributes.iterator(), null));
-        writer.add(factory.createCharacters(String.valueOf(content)));
-        writer.add(factory.createEndElement(QName.valueOf(tag), null));
-    }
-
-    private void writeTag(String tag, double content) throws XMLStreamException {
-        writer.add(factory.createStartElement(QName.valueOf(tag), null, null));
-        writer.add(factory.createCharacters(String.valueOf(content)));
-        writer.add(factory.createEndElement(QName.valueOf(tag), null));
-    }
-
-    private void writeTag(String tag, String content) throws XMLStreamException {
-        writer.add(factory.createStartElement(QName.valueOf(tag), null, null));
-        writer.add(factory.createCharacters(content));
-        writer.add(factory.createEndElement(QName.valueOf(tag), null));
-    }
-
-    private void writeTag(String tag, String content, List<Attribute> attributes) throws XMLStreamException {
-        writer.add(factory.createStartElement(QName.valueOf(tag), attributes.iterator(), null));
-        writer.add(factory.createCharacters(content));
-        writer.add(factory.createEndElement(QName.valueOf(tag), null));
-    }
-
-    private void writeTag(String tag, List<Attribute> attributes, IContentWriter content) throws XMLStreamException {
-        writer.add(factory.createStartElement(QName.valueOf(tag), attributes.iterator(), null));
-        content.write();
-        writer.add(factory.createEndElement(QName.valueOf(tag), null));
-    }
-
-    interface IContentWriter {
-        void write() throws XMLStreamException;
     }
 }
